+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 # Changelog
 
+## 1.0.4 (\_\_\_)
+
+### New Features
+
+---
+
+- [WUPTimeObject](src/objects/timeObject.ts). Add init by date to constructor
+- [DateControl](src/controls/date.ts). Add sync with time (value, min, max)
+
+---
+
+---
+
 ## 1.0.3 (Dec 9, 2023)
 
 ### New Features